Refactor Home component (second pass)
Removes more duplicate logic from Home component by moving it out to a
reusable <Section /> component.
