+++ admin/client/App/screens/Home/index.js
@@ -9,6 +9,7 @@ import { connect } from 'react-redux';
 
 import { plural } from '../../../utils/string';
 import Lists from './components/Lists';
+import Section from './components/Section';
 import AlertMessages from '../../shared/AlertMessages';
 import {
 	loadCounts,
@@ -20,32 +21,6 @@ var HomeView = React.createClass({
 	// from the API
 	componentDidMount () {
 		this.props.dispatch(loadCounts());
-	},
-	// Certain section name have an icon associated with the for a nicer view
-	getHeadingIconClasses (navSectionKey) {
-		const icons = [
-			{ icon: 'book', sections: ['books', 'posts', 'blog', 'blog-posts', 'stories', 'news-stories', 'content'] },
-			{ icon: 'briefcase', sections: ['businesses', 'companies', 'listings', 'organizations', 'partners'] },
-			{ icon: 'calendar', sections: ['events', 'dates'] },
-			{ icon: 'clock', sections: ['classes', 'hours', 'times'] },
-			{ icon: 'file-media', sections: ['gallery', 'galleries', 'images', 'photos', 'pictures'] },
-			{ icon: 'file-text', sections: ['attachments', 'docs', 'documents', 'files'] },
-			{ icon: 'location', sections: ['locations', 'markers', 'places'] },
-			{ icon: 'mail', sections: ['emails', 'enquiries'] },
-			{ icon: 'megaphone', sections: ['broadcasts', 'jobs', 'talks'] },
-			{ icon: 'organization', sections: ['contacts', 'customers', 'groups', 'members', 'people', 'speakers', 'teams', 'users'] },
-			{ icon: 'package', sections: ['boxes', 'items', 'packages', 'parcels'] },
-			{ icon: 'tag', sections: ['tags'] },
-		];
-		const classes = icons
-			.filter(obj => obj.sections.indexOf(navSectionKey) !== -1)
-			.map(obj => `octicon-${obj.icon}`);
-
-		if (!classes.length) {
-			classes.push('octicon-primitive-dot');
-		}
-
-		return ['dashboard-group__heading-icon', 'octicon', ...classes].join(' ');
 	},
 	getCount (key) {
 		// If we have previous counts already saved, show them while we fetch
@@ -78,11 +53,7 @@ var HomeView = React.createClass({
 			<div>
 				{Keystone.nav.sections.map((navSection) => {
 					return (
-						<div className="dashboard-group" key={navSection.key} data-section-label={navSection.label}>
-							<div className="dashboard-group__heading">
-								<span className={this.getHeadingIconClasses(navSection.key)} />
-								{navSection.label}
-							</div>
+						<Section key={navSection.key} id={navSection.key} label={navSection.label}>
 							<div className="dashboard-group__lists">
 								<Lists
 									counts={this.props.counts}
@@ -90,7 +61,7 @@ var HomeView = React.createClass({
 									spinner={this.getSpinner()}
 								/>
 							</div>
-						</div>
+						</Section>
 					);
 				})}
 				{this.renderOrphanedLists()}
@@ -99,13 +70,8 @@ var HomeView = React.createClass({
 	},
 	renderOrphanedLists () {
 		if (!Keystone.orphanedLists.length) return;
-		let sectionLabel = 'Other';
 		return (
-			<div className="dashboard-group" data-section-label={sectionLabel}>
-				<div className="dashboard-group__heading">
-					<span className="dashboard-group__heading-icon octicon octicon-database" />
-					{sectionLabel}
-				</div>
+			<Section label="Other" iconClass="dashboard-group__heading-icon octicon octicon-database">
 				<div className="dashboard-group__lists">
 					<Lists
 						counts={this.props.counts}
@@ -113,7 +79,7 @@ var HomeView = React.createClass({
 						spinner={this.getSpinner()}
 					/>
 				</div>
-			</div>
+			</Section>
 		);
 	},
 	render () {
